move_getters: init with getters from tui and api
